Move TypeOrBoundsTreeOps into TreeOps
