+++ tests/run/tasty-custom-show/quoted_1.scala
@@ -44,7 +44,6 @@ object Macros {
       def showTree(tree: Tree)(implicit ctx: Context): String = "Tree"
       def showCaseDef(caseDef: CaseDef)(implicit ctx: Context): String = "CaseDef"
       def showPattern(pattern: Pattern)(implicit ctx: Context): String = "Pattern"
-      def showTypeOrBoundsTree(tpt: TypeOrBoundsTree)(implicit ctx: Context): String = "TypeOrBoundsTree"
       def showTypeOrBounds(tpe: TypeOrBounds)(implicit ctx: Context): String = "TypeOrBounds"
       def showConstant(const: Constant)(implicit ctx: Context): String = "Constant"
       def showSymbol(symbol: Symbol)(implicit ctx: Context): String = "Symbol"
